+++ src/commands/system/users.js
@@ -1,0 +1,7 @@
+const chalk = require("chalk");
+
+const users = (...args) => {
+  console.log(chalk.italic("Command functionality not available.\n"));
+};
+
+module.exports = users;